a53x: overlay: Disable framework dimming
This morning I woke up to my device somehow not displaying anything, but ringing
my alarm sound to its maximum. That's because I don't use Adaptive Brightness
and it seems like if you manually set the brightness to minimum (as you should
do before you go to sleep), it somehow dims the lock screen to this point that
almost nothing is visible.

It's also annoying, because I can see the difference/sudden switch between the lockscreen
brightness and normal (anything but lockscreen) brightness.

So, disable this. It doesn't affect anything either way.